Improved grid storage and cell attributes
Attributes (including color) are now stored in a dictionary.
grid._cells is now a dictionary of (xpos, ypos) => {attr_dict}.  Any
points on the grid that are not present in grid._cells are rendered
using the default color value.

Defaults are also handled differently, using a _defaults dictionary
which holds fallbacks for any cell attribute not present.

Comparison is now done using == instead of the custom equals for colors.