Grid uses NotifiableDict, minor drawing tweaks
NotifiableDict is used for cell attributes, so that we can force a
redraw when there's a default change.

Corrected bug in hit_check
Moved redraw logic to surface so that force_redraw can simply set the
redraw flag.  This also cuts down on the number of times we append the
full list of cells to the _dirty variable.