Add clip flag to hit_check, corrected __getitem__
The clip flag for hit check allows you to constrain checks to points on
the grid.  When the clip flag is false and use_nearest is true, points
outside the grid will resolve to the nearest point on the grid.

__getitem__ was using slice.indices, which works for almost all cases -
except when stop is -1, which it wraps to the top value.  This is bad
for negative slices, which are used in multi_select.