Cleaned up multi-select demo
The logic for updating the selection is much clearer now, and actually
runs quite a bit faster.

Also updated the event loop to only check on mouse movement.

One further optimization that's not really needed would be to track the
endpoint as well as the selection, and only try to update the selection
when the endpoints are different (so we avoid the set comparison of the
selections)