Added Color class
Color supports swizzling and indexing, with a default ordering of rgba.

swizzling means you can access the colors in any order, in case you want
blue then alpha then green and no red, you can use my_color.bag and get
a list of length three back.

grid is going to get a bit of an overhaul in the way of a pluggable
renderer.  This will be a pretty big overhaul, but afterwards it will be
easy for people to plug their own rendering in instead of being forced
to use color.  color will either be an available renderer, or the
default, so that it's easy for people to get started with grid.